type size fixes, use unordered_map

+++ chapter01/02_vulkan_assimp/vulkan/Texture.cpp
@@ -11,7 +11,7 @@ bool Texture::loadTexture(VkRenderData &renderData, VkTextureData &texData, std:
   int texWidth;
   int texHeight;
   int numberOfChannels;
-  int mipmapLevels = 1;
+  uint32_t mipmapLevels = 1;
 
   stbi_set_flip_vertically_on_load(flipImage);
   /* always load as RGBA */
@@ -53,7 +53,7 @@ bool Texture::loadTexture(VkRenderData &renderData, VkTextureData &texData, std:
     Logger::log(1, "%s error: could not map texture memory (error: %i)\n", __FUNCTION__, result);
     return false;
   }
-  std::memcpy(uploadData, textureData, static_cast<uint32_t>(imageSize));
+  std::memcpy(uploadData, textureData, imageSize);
   vmaUnmapMemory(renderData.rdAllocator, stagingBufferAlloc);
   vmaFlushAllocation(renderData.rdAllocator, stagingBufferAlloc, 0, imageSize);
 
@@ -79,7 +79,7 @@ bool Texture::loadTexture(VkRenderData& renderData, VkTextureData& texData, std:
   int texWidth;
   int texHeight;
   int numberOfChannels;
-  int mipmapLevels = 1;
+  uint32_t mipmapLevels = 1;
 
   /* allow to flip the image, similar to file loaded from disk */
   stbi_set_flip_vertically_on_load(flipImage);
@@ -129,7 +129,7 @@ bool Texture::loadTexture(VkRenderData& renderData, VkTextureData& texData, std:
     Logger::log(1, "%s error: could not map texture memory (error: %i)\n", __FUNCTION__, result);
     return false;
   }
-  std::memcpy(uploadData, data, static_cast<uint32_t>(imageSize));
+  std::memcpy(uploadData, data, imageSize);
   vmaUnmapMemory(renderData.rdAllocator, stagingBufferAlloc);
   vmaFlushAllocation(renderData.rdAllocator, stagingBufferAlloc, 0, imageSize);
 
@@ -146,15 +146,15 @@ bool Texture::loadTexture(VkRenderData& renderData, VkTextureData& texData, std:
 }
 
 bool Texture::uploadToGPU(VkRenderData& renderData, VkTextureData& texData, VkTextureStagingBuffer &stagingData,
-      int width, int height, bool generateMipmaps, int mipmapLevels) {
+      uint32_t width, uint32_t height, bool generateMipmaps, uint32_t mipmapLevels) {
   /* upload */
   VkCommandBuffer uploadCommandBuffer = CommandBuffer::createSingleShotBuffer(renderData);
 
   VkImageCreateInfo imageInfo{};
   imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
   imageInfo.imageType = VK_IMAGE_TYPE_2D;
-  imageInfo.extent.width = static_cast<uint32_t>(width);
-  imageInfo.extent.height = static_cast<uint32_t>(height);
+  imageInfo.extent.width = width;
+  imageInfo.extent.height = height;
   imageInfo.extent.depth = 1;
   imageInfo.mipLevels = mipmapLevels;
   imageInfo.arrayLayers = 1;
@@ -192,8 +192,8 @@ bool Texture::uploadToGPU(VkRenderData& renderData, VkTextureData& texData, VkTe
   stagingBufferTransferBarrier.dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
 
   VkExtent3D textureExtent{};
-  textureExtent.width = static_cast<uint32_t>(width);
-  textureExtent.height = static_cast<uint32_t>(height);
+  textureExtent.width = width;
+  textureExtent.height = height;
   textureExtent.depth = 1;
 
   VkBufferImageCopy stagingBufferCopy{};
